BJS2-78003. Unit-тесты для задачи с подписками - PR правки
